[ANDROSDK-1994] apply PR comments
